Move pickle tests
